Revert "New method aded"
This reverts commit 9179e09f787d4b20d25014120e4d3fd60aabd089.

+++ Auction.BusinessLogic/Services/CategoryService.cs
@@ -34,16 +34,6 @@ namespace Auction.BusinessLogic.Services
             return Adapter.Adapt<CategoryDTO>(category);
         }
 
-        public TradingLotDTO GetLotFromCategory(int categoryId, int lotId)
-        {
-            var category = GetCategoryById(categoryId);
-
-            var lotDto = category.TradingLots.FirstOrDefault(lot => lot.Id == lotId)
-                ?? throw new NotFoundException($"Lot with id: {lotId} in {category.Name}");
-
-            return lotDto;
-        }
-
         public IEnumerable<CategoryDTO> GetAllCategories()
         {
             return Adapter.Adapt<List<CategoryDTO>>(Database.Categories.FindCategories());
@@ -52,77 +42,49 @@ namespace Auction.BusinessLogic.Services
         public void RemoveCategoryById(int id)
         {
             if (id == 1)
-                throw new AuctionException("You can`t delete default category");
+                throw new AuctionException("You cant delete default category");
 
-            var categoryToDelete = Database.Categories.GetCategoryById(id)
-                ?? throw new NotFoundException($"Category with id: {id}");
+            var categoryToDelete = GetCategoryById(id);
+
+            if (categoryToDelete == null)
+                throw new NullReferenceException();
 
             if (categoryToDelete.TradingLots != null)
-                MoveLotsToCategory(categoryToDelete.TradingLots);
+               // MoveLotsToDefaultCategory(categoryToDelete.TradingLots);
 
             Database.Categories.DeleteCategoryById(categoryToDelete.Id);
             Database.Save();
         }
 
-        private void MoveLotsToCategory(ICollection<TradingLot> tradingLots, int categoryId = 1)
+        private void MoveLotsToDefaultCategory(ICollection<TradingLot> tradingLots, int defaultId = 1)
         {
-            var defaultCat = Database.Categories.GetCategoryById(categoryId);
-            tradingLots.ToList().ForEach(lot => {
-                lot.Category = defaultCat;
-                lot.CategoryId = defaultCat.Id;
-                Database.TradingLots.UpdadeTradingLot(lot);
-            } );
+            var defaultCat = Database.Categories.GetCategoryById(defaultId);
+            tradingLots.ToList().ForEach(lot => lot.Category = defaultCat);
         }
-
-        public void EditCategory(int id, CategoryDTO category)
+        //maybe change name to ChangeCategoryName
+        public void EditCategory(CategoryDTO category)
         {
             if (category == null)
                 throw new ArgumentNullException(nameof(category));
-            var temp = Database.Categories.GetCategoryById(id) 
-                ?? throw new NotFoundException($"Category with id: {id}");
 
-            //temp.Name = category.Name;
-            var updatedCategory = Adapter.Adapt<Category>(category);
+            var temp = Database.Categories.GetCategoryById(category.Id);
 
-            var linq = from dto in category.TradingLots
-                       join db in temp.TradingLots on dto.Id equals db.Id
-                       select Adapter.Adapt<TradingLot>(db);
+            if (temp == null)
+                throw new ArgumentNullException(nameof(temp));
 
-            updatedCategory.TradingLots = linq.ToList();
-
-            Database.Categories.UpdadeCategory(updatedCategory);
-            Database.Save();
-        }
-
-
-        public void ChangeCategoryName(int id, string name)
-        {
-            var temp = Database.Categories.GetCategoryById(id)
-                ?? throw new NotFoundException($"Category with id: {id}");
-            if(!IsCategoryNameFree(name))
-                throw new AuctionException($"Category {name} already exists.");
-
-            temp.Name = name;
+            temp.Name = category.Name;
 
             Database.Categories.UpdadeCategory(temp);
             Database.Save();
         }
-
-        //check is category with such name is already exists and add constraint to category name as unique!!!
+        //check is category with such name is already exists
         public void CreateCategory(CategoryDTO category)
         {
             if (category == null)
                 throw new ArgumentNullException(nameof(category));
-            if (!IsCategoryNameFree(category.Name))
-                throw new AuctionException($"Category {category.Name} already exists.");
 
             Database.Categories.AddCategory(new Category { Name = category.Name });
             Database.Save();
-        }
-        //maybe change method equeals to more specified
-        private bool IsCategoryNameFree(string name)
-        {
-            return Database.Categories.Entities.FirstOrDefault(category => category.Name.Equals(name)) is null;
         }
     }
 }
